fix #62, sizes no longer a plugin
