+++ tests/test_7_gaps_sizes.py
@@ -1,0 +1,28 @@
+from tempfile import NamedTemporaryFile
+from genomepy.utils import generate_gap_bed, generate_fa_sizes
+
+
+def test_gaps():
+    infa = "tests/data/gap.fa"
+    outbed = "tests/data/gap.bed"
+
+    tmp = NamedTemporaryFile().name
+    generate_gap_bed(infa, tmp)
+
+    result = open(tmp).read()
+    expect = open(outbed).read()
+
+    assert result == expect
+
+
+def test_sizes():
+    infa = "tests/data/gap.fa"
+
+    tmp = NamedTemporaryFile().name
+    generate_fa_sizes(infa, tmp)
+
+    result = open(tmp).read()
+
+    assert result == "chr1\t28\n" \
+                     "chr2\t45\n" \
+                     "chr3\t15\n"
